pyproject.toml
[tool.poetry]
name = "test-task-career-centr"
version = "0.1.0"
description = ""
authors = ["aisaev <an.isaev7@gmail.com>"]
readme = "README.md"
packages = [{include = "test_task_career_centr"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
django-environ = "^0.10.0"
django-filter = "^23.2"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.26.2"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"
pre-commit = "^3.3.2"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-lenght = 120
target-version = ["py310"]
skip-string-normalization = true
skip-magic-trailing-coma = true
exclude = '''
(
  /(
      \.git
    | \.github
    | \.venv
  )/
)
'''
